+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## ChangeLog
 
+### 2.x.x
+* Added a function that returns all the elements that corresponding with a tag name
+
 ### 2.0.3
 * Added support to input disabled
 * Added support for `select` element